requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
defusedxml==0.7.1
Django==4.1.5
djangorestframework==3.14.0
httpie==3.2.1
idna==3.4
Markdown==3.4.1
multidict==6.0.4
Pygments==2.14.0
PySocks==1.7.1
pytz==2022.7
requests==2.28.1
requests-toolbelt==0.10.1
rich==13.0.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
